making sense out of the chaos that is the lib directory - round 1
