+++ lib/fix_spec/matchers/have_fix_path.rb
@@ -2,6 +2,7 @@ module FIXSpec
   module Matchers
     class HaveFIXPath
       attr_reader :path
+
       def initialize path
         @path = path
         @json_matcher = JsonSpec::Matchers::HaveJsonPath.new path
